+++ test/Compze.Tests.Integration/InProcess/Given_an_endpoint_declaring_only_in_process_tessaging_and_typermedia.cs
@@ -1,10 +1,10 @@
 ﻿using Compze.Abstractions.Hosting.Public;
+using Compze.Abstractions.Tessaging.Public;
 using Compze.DependencyInjection;
 using Compze.DependencyInjection.Abstractions;
 using Compze.Hosting.Testing;
 using Compze.Internals.SystemCE.ThreadingCE.TasksCE;
 using Compze.Must;
-using Compze.Tessaging;
 using Compze.Tessaging.Abstractions.Tessaging.Hosting.TessageHandling.Registration.Public;
 using Compze.Tessaging.Hosting;
 using Compze.Tests.Infrastructure;
@@ -62,7 +62,7 @@ public class Given_an_endpoint_declaring_only_in_process_tessaging_and_typermedi
    [PCT] public void a_tevent_published_in_process_reaches_the_endpoints_subscriber()
    {
       _endpoint.ServiceLocator.Resolve<IScopeFactory>().ExecuteInIsolatedScope(scope =>
-         scope.Resolve<IInProcessTeventPublisher>().Publish(new MySpecialGreetingRequestedTevent(), scope));
+         scope.Resolve<ITeventPublisher>().Publish(new MySpecialGreetingRequestedTevent()));
       _receivedTevents.Must().HaveCount(1);
    }
 }